refactor(cloud-mode): cleanup cloud platform UI, API structure and imports
- Remove deprecated `cloud-mode-ky.ts` and consolidate API clients.
- Move `createInstance` API back to the frontend and refine its implementation.
- Clean up `cloud-platform.svelte` and `cloud-mode-panel.svelte` by removing extensive test sections and unused imports.
- Update `cloud-env-state.svelte.ts` to use the correct API imports.
